backend/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
bcrypt==4.1.1
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
fakeredis==2.21.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-RESTful==0.3.10
flask-sock==0.7.0
Flask-SocketIO==5.3.6
h11==0.14.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mongoengine==0.27.0
mongomock==4.1.2
ordered-set==4.1.0
packaging==23.2
pluggy==1.3.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.4
pytest-order==1.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
rich==13.7.0
sentinels==1.0.0
simple-websocket==1.0.0
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.9.0
tzlocal==5.2
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
